backend/requirements.txt
# Django and REST Framework
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.1

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.0.2
pytest-django==4.8.0

# Code quality
black==24.2.0
flake8==7.0.0
